Added support for send simulate, changed recv-loop
Added the new function send_simulate.
irsend (the LIRC tool) have a mode called simulate where it could simulate received commands. This is useful for sending commands to LIRC.
The syntax looks like this:
	r = lirc_command_init(&cmd, "SIMULATE %016x %02x %s %s\n", scancode, repeat, keysym, remote);

Also changed the recv-loop to look for "BEGIN" in the whole data-string.
The first characters in data isn't always "BEGIN", it can contain newlines and other messages.